requirements.txt
numpy~=1.21.4
opencv-python~=4.2.0.34
pyyaml~=5.3.1
